+++ sdk/src/main/java/ly/count/android/sdk/ModuleConfiguration.java
@@ -2,7 +2,12 @@ package ly.count.android.sdk;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import java.util.HashSet;
 import java.util.Iterator;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -47,6 +52,17 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
     final static String keyRBOMRQPercentage = "bom_rqp";
     final static String keyRBOMRequestAge = "bom_ra";
     final static String keyRBOMDuration = "bom_d";
+    final static String keyRUserPropertyCacheLimit = "upcl";
+    final static String keyREventBlacklist = "eb";
+    final static String keyRUserPropertyBlacklist = "upb";
+    final static String keyRSegmentationBlacklist = "sb";
+    final static String keyREventSegmentationBlacklist = "esb"; // json
+    final static String keyREventWhitelist = "ew";
+    final static String keyRUserPropertyWhitelist = "upw";
+    final static String keyRSegmentationWhitelist = "sw";
+    final static String keyREventSegmentationWhitelist = "esw"; // json
+    final static String keyRJourneyTriggerEvents = "jte";
+
     // FLAGS
     boolean currentVTracking = true;
     boolean currentVNetworking = true;
@@ -64,6 +80,14 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
     double currentVBOMRQPercentage = 0.5;
     int currentVBOMRequestAge = 24; // in hours
     int currentVBOMDuration = 60; // in seconds
+    int currentVUserPropertyCacheLimit = 100;
+
+    // FILTERS
+    FilterList<Set<String>> currentVEventFilterList = new FilterList<>(new HashSet<>(), false);
+    FilterList<Set<String>> currentVUserPropertyFilterList = new FilterList<>(new HashSet<>(), false);
+    FilterList<Set<String>> currentVSegmentationFilterList = new FilterList<>(new HashSet<>(), false);
+    FilterList<Map<String, Set<String>>> currentVEventSegmentationFilterList = new FilterList<>(new ConcurrentHashMap<>(), false);
+    Set<String> currentVJourneyTriggerEvents = new HashSet<>();
 
     // SERVER CONFIGURATION PARAMS
     Integer serverConfigUpdateInterval; // in hours
@@ -207,6 +231,7 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
         currentVBOMRQPercentage = extractValue(keyRBOMRQPercentage, sb, currentVBOMRQPercentage, currentVBOMRQPercentage, Double.class, (Double value) -> value > 0.0 && value < 1.0);
         currentVBOMRequestAge = extractValue(keyRBOMRequestAge, sb, currentVBOMRequestAge, currentVBOMRequestAge, Integer.class, (Integer value) -> value > 0);
         currentVBOMDuration = extractValue(keyRBOMDuration, sb, currentVBOMDuration, currentVBOMDuration, Integer.class, (Integer value) -> value > 0);
+        currentVUserPropertyCacheLimit = extractValue(keyRUserPropertyCacheLimit, sb, currentVUserPropertyCacheLimit, currentVUserPropertyCacheLimit, Integer.class, (Integer value) -> value > 0);
 
         clyConfig.setMaxRequestQueueSize(extractValue(keyRReqQueueSize, sb, clyConfig.maxRequestQueueSize, clyConfig.maxRequestQueueSize, Integer.class, (Integer value) -> value > 0));
         clyConfig.setEventQueueSizeToSend(extractValue(keyREventQueueSize, sb, clyConfig.eventQueueSizeThreshold, Countly.sharedInstance().EVENT_QUEUE_SIZE_THRESHOLD, Integer.class, (Integer value) -> value > 0));
@@ -222,10 +247,106 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
         clyConfig.setRequiresConsent(extractValue(keyRConsentRequired, sb, clyConfig.shouldRequireConsent, clyConfig.shouldRequireConsent));
         clyConfig.setRequestDropAgeHours(extractValue(keyRDropOldRequestTime, sb, clyConfig.dropAgeHours, clyConfig.dropAgeHours, Integer.class, (Integer value) -> value >= 0));
 
+        updateListingFilters();
+
         String updatedValues = sb.toString();
         if (!updatedValues.isEmpty()) {
             L.i("[ModuleConfiguration] updateConfigVariables, SDK configuration has changed, notifying the SDK, new values: [" + updatedValues + "]");
             _cly.onSdkConfigurationChanged(clyConfig);
+        }
+    }
+
+    private void updateListingFilters() {
+        L.d("[ModuleConfiguration] updateListingFilters, current listing filters before updating: \n" +
+            "Event Filter List: " + currentVEventFilterList.filterList + ", isWhitelist: " + currentVEventFilterList.isWhitelist + "\n" +
+            "User Property Filter List: " + currentVUserPropertyFilterList.filterList + ", isWhitelist: " + currentVUserPropertyFilterList.isWhitelist + "\n" +
+            "Segmentation Filter List: " + currentVSegmentationFilterList.filterList + ", isWhitelist: " + currentVSegmentationFilterList.isWhitelist + "\n" +
+            "Event Segmentation Filter List: " + currentVEventSegmentationFilterList.filterList + ", isWhitelist: " + currentVEventSegmentationFilterList.isWhitelist + "\n" +
+            "Journey Trigger Events: " + currentVJourneyTriggerEvents);
+        JSONArray eventBlacklistJSARR = latestRetrievedConfiguration.optJSONArray(keyREventBlacklist);
+        JSONArray eventWhitelistJSARR = latestRetrievedConfiguration.optJSONArray(keyREventWhitelist);
+        JSONArray userPropertyBlacklistJSARR = latestRetrievedConfiguration.optJSONArray(keyRUserPropertyBlacklist);
+        JSONArray userPropertyWhitelistJSARR = latestRetrievedConfiguration.optJSONArray(keyRUserPropertyWhitelist);
+        JSONArray segmentationBlacklistJSARR = latestRetrievedConfiguration.optJSONArray(keyRSegmentationBlacklist);
+        JSONArray segmentationWhitelistJSARR = latestRetrievedConfiguration.optJSONArray(keyRSegmentationWhitelist);
+        JSONObject eventSegmentationBlacklistJSOBJ = latestRetrievedConfiguration.optJSONObject(keyREventSegmentationBlacklist);
+        JSONObject eventSegmentationWhitelistJSOBJ = latestRetrievedConfiguration.optJSONObject(keyREventSegmentationWhitelist);
+        JSONArray journeyTriggerEventsJSARR = latestRetrievedConfiguration.optJSONArray(keyRJourneyTriggerEvents);
+
+        if (eventBlacklistJSARR != null) {
+            extractFilterSetFromJSONArray(eventBlacklistJSARR, currentVEventFilterList.filterList);
+            currentVEventFilterList.isWhitelist = false;
+        } else if (eventWhitelistJSARR != null) {
+            extractFilterSetFromJSONArray(eventWhitelistJSARR, currentVEventFilterList.filterList);
+            currentVEventFilterList.isWhitelist = true;
+        }
+
+        if (userPropertyBlacklistJSARR != null) {
+            extractFilterSetFromJSONArray(userPropertyBlacklistJSARR, currentVUserPropertyFilterList.filterList);
+            currentVUserPropertyFilterList.isWhitelist = false;
+        } else if (userPropertyWhitelistJSARR != null) {
+            extractFilterSetFromJSONArray(userPropertyWhitelistJSARR, currentVUserPropertyFilterList.filterList);
+            currentVUserPropertyFilterList.isWhitelist = true;
+        }
+
+        if (segmentationBlacklistJSARR != null) {
+            extractFilterSetFromJSONArray(segmentationBlacklistJSARR, currentVSegmentationFilterList.filterList);
+            currentVSegmentationFilterList.isWhitelist = false;
+        } else if (segmentationWhitelistJSARR != null) {
+            extractFilterSetFromJSONArray(segmentationWhitelistJSARR, currentVSegmentationFilterList.filterList);
+            currentVSegmentationFilterList.isWhitelist = true;
+        }
+
+        if (eventSegmentationBlacklistJSOBJ != null) {
+            currentVEventSegmentationFilterList.filterList.clear();
+            currentVEventSegmentationFilterList.isWhitelist = false;
+            Iterator<String> keys = eventSegmentationBlacklistJSOBJ.keys();
+            while (keys.hasNext()) {
+                String key = keys.next();
+                JSONArray jsonArray = eventSegmentationBlacklistJSOBJ.optJSONArray(key);
+                if (jsonArray != null) {
+                    Set<String> filterSet = new HashSet<>();
+                    extractFilterSetFromJSONArray(jsonArray, filterSet);
+                    currentVEventSegmentationFilterList.filterList.put(key, filterSet);
+                }
+            }
+        } else if (eventSegmentationWhitelistJSOBJ != null) {
+            currentVEventSegmentationFilterList.filterList.clear();
+            currentVEventSegmentationFilterList.isWhitelist = true;
+            Iterator<String> keys = eventSegmentationWhitelistJSOBJ.keys();
+            while (keys.hasNext()) {
+                String key = keys.next();
+                JSONArray jsonArray = eventSegmentationWhitelistJSOBJ.optJSONArray(key);
+                if (jsonArray != null) {
+                    Set<String> filterSet = new HashSet<>();
+                    extractFilterSetFromJSONArray(jsonArray, filterSet);
+                    currentVEventSegmentationFilterList.filterList.put(key, filterSet);
+                }
+            }
+        }
+
+        if (journeyTriggerEventsJSARR != null) {
+            extractFilterSetFromJSONArray(journeyTriggerEventsJSARR, currentVJourneyTriggerEvents);
+        }
+
+        L.d("[ModuleConfiguration] updateListingFilters, current listing filters after updating: \n" +
+            "Event Filter List: " + currentVEventFilterList.filterList + ", isWhitelist: " + currentVEventFilterList.isWhitelist + "\n" +
+            "User Property Filter List: " + currentVUserPropertyFilterList.filterList + ", isWhitelist: " + currentVUserPropertyFilterList.isWhitelist + "\n" +
+            "Segmentation Filter List: " + currentVSegmentationFilterList.filterList + ", isWhitelist: " + currentVSegmentationFilterList.isWhitelist + "\n" +
+            "Event Segmentation Filter List: " + currentVEventSegmentationFilterList.filterList + ", isWhitelist: " + currentVEventSegmentationFilterList.isWhitelist + "\n" +
+            "Journey Trigger Events: " + currentVJourneyTriggerEvents);
+    }
+
+    private void extractFilterSetFromJSONArray(@Nullable JSONArray jsonArray, @NonNull Set<String> targetSet) {
+        if (jsonArray == null) {
+            return;
+        }
+        targetSet.clear();
+        for (int i = 0; i < jsonArray.length(); i++) {
+            String item = jsonArray.optString(i, null);
+            if (item != null) {
+                targetSet.add(item);
+            }
         }
     }
 
@@ -293,6 +414,7 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
                 case keyRLimitBreadcrumb:
                 case keyRLimitTraceLine:
                 case keyRLimitTraceLength:
+                case keyRUserPropertyCacheLimit:
                     isValid = value instanceof Integer && ((Integer) value) > 0;
                     break;
 
@@ -309,6 +431,21 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
                 // --- Double between 0.0 and 1.0 ---
                 case keyRBOMRQPercentage:
                     isValid = value instanceof Double && ((Double) value > 0.0 && (Double) value < 1.0);
+                    break;
+
+                // --- Filtering keys ---
+                case keyREventBlacklist:
+                case keyRSegmentationBlacklist:
+                case keyRUserPropertyBlacklist:
+                case keyREventWhitelist:
+                case keyRSegmentationWhitelist:
+                case keyRUserPropertyWhitelist:
+                case keyRJourneyTriggerEvents:
+                    isValid = value instanceof JSONArray;
+                    break;
+                case keyREventSegmentationBlacklist:
+                case keyREventSegmentationWhitelist:
+                    isValid = value instanceof JSONObject;
                     break;
                 // --- Unknown keys ---
                 default:
@@ -353,6 +490,8 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
             L.w("[ModuleConfiguration] saveAndStoreDownloadedConfig, Failed to merge version/timestamp.", e);
         }
 
+        removeListingFilterKeysFromConfig(newInner);
+
         Iterator<String> keys = newInner.keys();
         while (keys.hasNext()) {
             String key = keys.next();
@@ -368,6 +507,34 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
 
         // Save updated config
         storageProvider.setServerConfig(latestRetrievedConfigurationFull.toString());
+    }
+
+    private void removeListingFilterKeysFromConfig(JSONObject newConfig) {
+        boolean hasAnyWhitelist = newConfig.has(keyREventWhitelist)
+            || newConfig.has(keyRUserPropertyWhitelist)
+            || newConfig.has(keyRSegmentationWhitelist)
+            || newConfig.has(keyREventSegmentationWhitelist);
+
+        boolean hasAnyBlacklist = newConfig.has(keyREventBlacklist)
+            || newConfig.has(keyRUserPropertyBlacklist)
+            || newConfig.has(keyRSegmentationBlacklist)
+            || newConfig.has(keyREventSegmentationBlacklist);
+
+        // Only remove opposite type when we actually have data for current type
+        if (hasAnyWhitelist) {
+            latestRetrievedConfiguration.remove(keyREventBlacklist);
+            latestRetrievedConfiguration.remove(keyRUserPropertyBlacklist);
+            latestRetrievedConfiguration.remove(keyRSegmentationBlacklist);
+            latestRetrievedConfiguration.remove(keyREventSegmentationBlacklist);
+        }
+
+        if (hasAnyBlacklist) {
+            latestRetrievedConfiguration.remove(keyREventWhitelist);
+            latestRetrievedConfiguration.remove(keyRUserPropertyWhitelist);
+            latestRetrievedConfiguration.remove(keyRSegmentationWhitelist);
+            latestRetrievedConfiguration.remove(keyREventSegmentationWhitelist);
+        }
+        // If neither has data, don't remove anything - preserve existing filters
     }
 
     /**
@@ -497,5 +664,29 @@ class ModuleConfiguration extends ModuleBase implements ConfigurationProvider {
 
     @Override public int getRequestTimeoutDurationMillis() {
         return _cly.config_.requestTimeoutDuration * 1000;
+    }
+
+    @Override public int getUserPropertyCacheLimit() {
+        return currentVUserPropertyCacheLimit;
+    }
+
+    @Override public FilterList<Set<String>> getEventFilterList() {
+        return currentVEventFilterList;
+    }
+
+    @Override public FilterList<Set<String>> getUserPropertyFilterList() {
+        return currentVUserPropertyFilterList;
+    }
+
+    @Override public FilterList<Set<String>> getSegmentationFilterList() {
+        return currentVSegmentationFilterList;
+    }
+
+    @Override public FilterList<Map<String, Set<String>>> getEventSegmentationFilterList() {
+        return currentVEventSegmentationFilterList;
+    }
+
+    @Override public Set<String> getJourneyTriggerEvents() {
+        return currentVJourneyTriggerEvents;
     }
 }